76007 restructured the pos to facilitate the aio AB#76007 AB#76019
