+++ UKHO.PeriodicOutputService/UKHO.PeriodicOutputService.Common/Configuration/EssApiConfiguration.cs
@@ -1,6 +1,9 @@
-﻿namespace UKHO.PeriodicOutputService.Fulfilment.Configuration
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace UKHO.PeriodicOutputService.Common.Configuration
 {
-    public interface IEssApiConfiguration
+    [ExcludeFromCodeCoverage]
+    public class EssApiConfiguration : IEssApiConfiguration
     {
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public string EssClientId { get; set; }
