+++ UKHO.PeriodicOutputService/UKHO.PeriodicOutputService.Common/Configuration/FssApiConfiguration.cs
@@ -1,6 +1,9 @@
-﻿namespace UKHO.PeriodicOutputService.Fulfilment.Configuration
+﻿using System.Diagnostics.CodeAnalysis;
+
+namespace UKHO.PeriodicOutputService.Common.Configuration
 {
-    public interface IFssApiConfiguration
+    [ExcludeFromCodeCoverage]
+    public class FssApiConfiguration : IFssApiConfiguration
     {
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public string FssClientId { get; set; }
@@ -9,7 +12,9 @@
         public string BatchStatusPollingDelayTime { get; set; }
         public int BlockSizeInMultipleOfKBs { get; set; }
         public int ParallelUploadThreadCount { get; set; }
+        public string BusinessUnit { get; set; }
+        public string PosReadUsers { get; set; }
+        public string PosReadGroups { get; set; }
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
-
     }
 }
