+++ UKHO.PeriodicOutputService/UKHO.PeriodicOutputService.Common/Configuration/IEssApiConfiguration.cs
@@ -1,9 +1,6 @@
-﻿using System.Diagnostics.CodeAnalysis;
-
-namespace UKHO.PeriodicOutputService.Fulfilment.Configuration
+﻿namespace UKHO.PeriodicOutputService.Common.Configuration
 {
-    [ExcludeFromCodeCoverage]
-    public class EssApiConfiguration : IEssApiConfiguration
+    public interface IEssApiConfiguration
     {
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public string EssClientId { get; set; }
