+++ UKHO.PeriodicOutputService/UKHO.PeriodicOutputService.Common/Configuration/IFssApiConfiguration.cs
@@ -1,9 +1,6 @@
-﻿using System.Diagnostics.CodeAnalysis;
-
-namespace UKHO.PeriodicOutputService.Fulfilment.Configuration
+﻿namespace UKHO.PeriodicOutputService.Common.Configuration
 {
-    [ExcludeFromCodeCoverage]
-    public class FssApiConfiguration : IFssApiConfiguration
+    public interface IFssApiConfiguration
     {
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
         public string FssClientId { get; set; }
@@ -12,9 +9,7 @@ namespace UKHO.PeriodicOutputService.Fulfilment.Configuration
         public string BatchStatusPollingDelayTime { get; set; }
         public int BlockSizeInMultipleOfKBs { get; set; }
         public int ParallelUploadThreadCount { get; set; }
-        public string BusinessUnit { get; set; }
-        public string PosReadUsers { get; set; }
-        public string PosReadGroups { get; set; }
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+
     }
 }
